@@ -1,14 +1,34 @@
+# == Schema Information
+#
+# Table name: leagues
+#
+#  id          :integer          not null, primary key
+#  name        :string(255)      not null
+#  start_at    :datetime         default(2014-10-24 15:06:30 +0000), not null
+#  end_at      :datetime         default(2014-10-24 15:06:30 +0000), not null
+#  limit_score :float            default(0.0), not null
+#  is_analy    :boolean          default(FALSE), not null
+#  data_dir    :string(255)      default(""), not null
+#  rule_file   :string(255)      default(""), not null
+#  is_active   :boolean          default(TRUE), not null
+#  created_at  :datetime
+#  updated_at  :datetime
+#
+
 # Read about factories at https://github.com/thoughtbot/factory_girl
 
 FactoryGirl.define do
   factory :league do
-    name "MyString"
-    start_at "2014-09-15"
-    end_at "2014-09-15"
-    limit_score 1.5
+    name 'テストリーグ'
+    start_at '2014-09-15'
+    end_at '2014-09-15'
+    limit_score 150.0
     is_analysis false
-    src_dir "MyString"
-    rule_file "MyString"
-    is_active false
+    is_active true
+
+    factory :before_submit do
+      src_dir ''
+      rule_file ''
+    end
   end
 end
